Skip headers for Not Modified fs responses
According to RFC 7232, 304 Not Modified responses should not generate
representation metadata other than the following fields: Cache-Control,
Content-Location, Date, ETag, Expires, and Vary.

This avoids a warning from hyper when logging is enabled:
"content-length value found, but empty body provided". While it is
allowed to provide Content-Length for Not Modified responses it is
unnecessary.