Add ability to send Ping ws Messages
This commit adds the ability to send WebSocket `Ping` messages.
This functionality is built into Tungstenite, but was not
previously exposed in Warp's public API.  Ultimately, it would be
helpful for Warp to expose functionality that automates sending
of pings, but this seems like a good first step and allows users
to manually send pings (for example, using a timer/loop) in the
meantime.