Emphasize that path-related filters should come first
To avoid the hard-to-debug errors that can result from including a filter to check the authorization header or body type first.
If a different filter type is included before the path, rejected requests from other routes will try matching against the given route.
The rejection will suggest that the request failed the header or body check for one route when it was actually a request for a completely different path.