A Result can be a Reply
Provide `impl Reply for Result<T, E>` where both `T` and `E`
implements Reply.  This impl makes no difference between an Ok and an
Err, but calls `into_response` for the contained data of either.  If
an error should be logged, that is the responsibility of the `Reply`
impl on the `E` type.