better API for CORS origin validation (#564)
Replace IntoOrigin with TryIntoOrigin, that returns TryIntoOriginError
upon receiving an invalid CORS Origin string. Make this trait and error
public instead of keeping them private. This gives the library users
better overall experience allowing to validate a CORS Origin string
before feeding it to Builder.

Since Builder expects TryIntoOrigin, which is implemented for Origin as
well, users can gracefully process the error when trying to convert a
string to Origin first, and then feed an Origin to the builder.

The builder API is backwards compatible. Feeding any &str to it is still
perfectly fine and it will panic upon meeting an invalid CORS Origin
value.