Expose the custom resolver to the warp interface
This allows users to inject their own CertResolver directly to rustls
for custom cert resolution.

Currently warp provides an interface for using a single certificate for
the duration of the lifecycle of a service.

In some cases, it's useful to be able to inject a custom resolver to
allow more granular use cases. For example, we want to inject a resolver
which periodically reloads the certs from disk to allow live renewals
without interrupting the service.

Exposing the CertResolver configuration allows any custom use case to be
implemented without having to account for every possibility within warp.